requirements.txt
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
dateparser==1.0.0
flake8==3.8.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pipdeptree==2.0.0
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pytest==6.2.3
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2021.1
qtmodern==0.2.0
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
service-identity==18.1.0
six==1.15.0
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
zope.interface==5.2.0
